+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 0.5.2
+- rework preventing redirects to non-https URLs available now in new Guzzle
+- add some unit tests for testing HTTPS to HTTP redirect
+
 ## 0.5.1
 - update `guzzlehttp/guzzle`
 - add all module dependencies to composer.json and spec file